\name{fia}
\alias{fia_obs}
\alias{fia_nul}
\docType{data}
\title{
Demo input species x site matrices
}
\description{
Data from referenced paper.
\code{fia_obs}{Observed data: Community matrix of abundance of tree species in Forest Inventory of America plots.}
\code{fia_nul}{Null data: Suitability score for tree species in locations of Forest Inventory of America plots, as predicted by MaxEnt regional models of species' distributions.}
}
\references{
Morueta-Holme, N., Blonder, B., et al. in preparation.
}