\name{netassoc-package}
\alias{netassoc-package}
\alias{netassoc}
\docType{package}
\title{
Inference of species assocations from co-occurrence data
}
\description{
Infers species associations from local and regional-scale co-occurrence data (species x site matrices). The resulting network can be analyzed using functions from the \code{igraph} network package.
}
\author{
Benjamin Blonder, Naia Morueta-Holme

Maintainer: Benjamin Blonder <bblonder@gmail.com>
}
\references{
Morueta-Holme, N., Blonder, B., et al. IN PREPARATION.
}
\keyword{ package }