\name{netassoc-package}
\alias{netassoc-package}
\alias{netassoc}
\docType{package}
\title{
Inference of Species Assocations from Co-Occurrence Data
}
\description{
Infers species associations from community matrices. Uses local and (optional) regional-scale co-occurrence data by comparing observed partial correlation coefficients between species to those estimated from regional species distributions. Extends Gaussian graphical models to a null modeling framework. Provides interface to a variety of inverse covariance matrix estimation methods.
}
\author{
Benjamin Blonder, Naia Morueta-Holme

Maintainer: Benjamin Blonder <bblonder@gmail.com>
}
\references{
Morueta-Holme, N., Blonder, B., et al. IN PREPARATION.
}
\keyword{ package }